A new read-hash extension has been added, for #\<, that makes the reader interpret expressions like #<φ> as (φ), where the count of '<' inside φ must equate the count of '>', unless the eof object occurs first.
Simultaneously, the id of instances of <goose> and its descendants
is no longer generated using gensym. Instead, it's its address,
as returned by the display function (reinterpreted as a number).
A minor inconvenience might be that the display procedure for GOOPS
objects displays a hex address, and by default the display procedure
uses decimal representation, so these two seem different to the human
eye. Fortunately, that's usually meaningless.