Finally, after many weeks of struggle,  subspaces distributed over the network have been added. The system is still far from perfection (there are still some issues with the chain-request macro, the notion of transactions hasn't been implemented properly (it would be perfect to make them work transparently; similarly, unresponded requests could be sent again after some specified time), and it's still questionable whether subspaces should be separate objects. Furthermore, there are some strange dependencies, as the 'subspace' module depends on the 'network' module, which seems illogical.
Among the architectural differences, the network-state-of function
and network-slot-value method (i.e. inheritable/extendable) were
added (the first one uses the second) that are used for transmitting
objects over the network. Likewise, it's not perfect, but it draws
certain idea for the future (hopefully -- more coherent and better
defined) architecture.

The display procedure in scm/3d-view.scm has been fixed, and also
slayer.scm is adjusted to that fix (it catches an exception that
is thrown when an invalid portal is to be displayed). Furthermore,
some debugging features have been assigned to the keys o,p,l.