Fixed some unresolved questions regarding informing clients about the change of state of an object. It's complicated, and regards the following question: if we have a class that is a descendant of <register-write-access>, we don't want to register write access when setting a slot that was introduced by the <register-write-access> class (or higher), or at the level, at which <register-write-access> was introduced.
To give the actual example, the <player> class inherited from
<network-object> and <3d-mesh>. <network-object> inherited
from <proto-network-object> and <register-write-access>, while
<3d-mesh> -- from <3d-shape>. This mean that (in the previous
solution) the accesses to slots of <3d-shape> were not recorded,
because that class was at the same level of inheritance as
<register-write-access>.
Now the behaviour has been modified, but it still isn't
satisfactory. The best solution would be to overload the
before-slot-set! for <network-object>, so that none of
its slots would be recorded.