Next step towards the implementation of chess demo: a separate module has been cut out (schess/elements.scm) which exports all the functions essential to implement schess; the chess.scm file (containing the rules) has been corrected slightly, and is now supported (albeit not entirely) by schess.scm demo.
Recent changes in the development of Schess haven't been tracked
by mercurial, so the process isn't documented properly. The current
version of Schess loads the initial board from chess.scm, displays
it and allows to move all the figures on the board freely. The next
point on the todo list is to allow only the moves specified by the
ruleset.